% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenRes-package.R
\docType{package}
\name{screenRes-package}
\alias{screenRes}
\alias{screenRes-package}
\title{screenRes: Get Screen Resolution}
\description{
Get the screen resolution for help in placing graphical components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaPrime7/screenRes}
  \item Report bugs at \url{https://github.com/AlphaPrime7/screenRes/issues}
}

}
\keyword{internal}
