% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenRes.R
\name{screen_res_dos_unix}
\alias{screen_res_dos_unix}
\alias{screen_res_win}
\title{Screen Resolution Dos and Unix}
\usage{
screen_res_dos_unix(screen_number = NULL)

screen_res_win()
}
\arguments{
\item{screen_number}{Number of Screens}
}
\value{
Screen Resolution

Screen Resolution
}
\description{
Screen Resolution Dos and Unix

Screen Resolution Windows
}
\examples{
\donttest{screen_res_dos_unix()}
\donttest{screen_res_win()}
}
\seealso{
[screen_res_dos_unix()]
}
