% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenRes-package.R
\docType{package}
\name{screenRes-package}
\alias{screenRes}
\alias{screenRes-package}
\title{screenRes: Get Screen Resolution}
\description{
Get the screen resolution for help in placing graphical components. The Package provides two system call functions that can facilitate obtaining the screen resolution for a screen or screens under scrutiny. The package is complementary to packages like tcltk and other packages that place GUI components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlphaPrime7/screenRes}
  \item Report bugs at \url{https://github.com/AlphaPrime7/screenRes/issues}
}

}
\author{
\strong{Maintainer}: Tingwei Adeck \email{awesome.tingwei@outlook.com} (\href{https://orcid.org/0009-0009-7450-8863}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Tesla Adeck [copyright holder]
  \item Amina Adeck [copyright holder]
}

}
\keyword{internal}
